Fixed many issues with the gun rendering. (#4)
* removed old changelog

* formatting

* added the ability to use smooth camera while holding a gun

* you can now hold guns in the left and right hand.

This mostly worked, but I found an issue with obfuscate. Perhaps you can use a copy of the original item stack when rendering to allow modification to said stack. That would allow for a lot more flexability when rendering. For example, you could render a model in the gui normally, then you could change the actual item that is rendered in third person or in first person when you want. That would allow the ability to clear an item when rendering when in reality, the item is still the same. Because of this issue I can not finish the rendering of the guns in third person. Look into what I asked for or maybe something different. By the way, the old render method was actually broken because it would render the same model for both items instead of a specific side. Please look into fixes for these issues.

* fixed any issues created in the last commit

I realized you can change the item renderered per hand once I did some digging.