Improved accuracy of weapons
The gun mod sends the clients current yaw and pitch to the server when shooting. Previously it was using the last yaw and pitch of the player and this would be inaccurate when playing is targeting a moving object. Although it now looks easy to write hacks, this was already possible before.